pipelines/ml/requirements.txt
kfp==2.12.1
google-cloud-aiplatform==1.82.0
google-cloud-bigquery==3.30.0
google-cloud-storage==2.19.0
pandas==2.2.3
db-dtypes==1.3.1
scikit-learn==1.6.1
joblib==1.4.2
